updated for compatibility with OpenRocket-23.09, restructed project, clean up
