+++ README.md
@@ -57,6 +57,11 @@ For most people jpype will be able to automatically find the JDK. However, if it
 - See `examples/` for usage examples
 - See [the OpenRocket wiki](https://github.com/openrocket/openrocket/wiki/Scripting-with-Python-and-JPype) for more info on usage and the examples 
 
+## Development
+- Use to install the package in editable mode:
+```
+pip install -e .
+```
 
 ## Credits
 - Richard Graham for the original script: [Source](https://sourceforge.net/p/openrocket/mailman/openrocket-devel/thread/4F17AA0C.1040002@rdg.cc/)
