+++ orhelper/__init__.py
@@ -1,7 +1,13 @@
-from ._orhelper import *
 from ._enums import *
+from .core.simulation_listener import *
+from .core.helper import *
+from .core.openrocket_instance import *
+from .core.jiterator import *
 
-__all__ =(
-    _orhelper.__all__ +
-    _enums.__all__
+__all__ = (
+    _enums.__all__ +
+    core.simulation_listener.__all__ +
+    core.helper.__all__ +
+    core.openrocket_instance.__all__ +
+    core.jiterator.__all__
 )
